fix chức năng xóa người dùng admin

+++ app/Http/Controllers/Admin/UserController.php
@@ -10,18 +10,21 @@ use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
-    public function index() {
+    public function index()
+    {
         $users = User::get();
         return view('admin.user.index', [
             'listUser' => $users
         ]);
     }
 
-    public function create() {
+    public function create()
+    {
         return view('admin.user.create');
     }
 
-    public function store(Request $request) {
+    public function store(Request $request)
+    {
         User::create([
             'name' => $request->name,
             'email' => $request->email,
@@ -31,14 +34,16 @@ class UserController extends Controller
         return redirect()->route('admin.user.index');
     }
 
-    public function edit($id) {
+    public function edit($id)
+    {
         $user = User::findOrFail($id);
         return view('admin.user.edit', [
             'user' => $user,
         ]);
     }
 
-    public function update(Request $request) {
+    public function update(Request $request)
+    {
         $user = User::findOrFail($request->id);
         $user->name = $request->name;
         $user->role = $request->role;
@@ -49,9 +54,19 @@ class UserController extends Controller
         return redirect()->route('admin.user.index');
     }
 
-    public function delete($id) {
+    public function delete($id)
+    {
         $user = User::findOrFail($id);
+
+        // Kiểm tra nếu người dùng có role là 0 (người quản trị)
+        if ($user->role == 0) {
+            // Trả về với thông báo lỗi
+            return redirect()->route('admin.user.index')->with('error', 'Không thể xóa người quản trị.');
+        }
+
+        // Xóa người dùng nếu không phải là người quản trị
         $user->delete();
-        return redirect()->route('admin.user.index');
+        return redirect()->route('admin.user.index')->with('success', 'Xóa người dùng thành công.');
     }
+
 }
